Fix SA double connection bug
New default behaviour if user tries to reconnect to an already open SA - close currently open SA (device handle 0) and re-open it, while throwing a warning to the user. As we only own one SA, and we know it is always assigned a device handle of 0, this workaround works. But it is very hacky, we need to think of something better...